added delete confirm model dialog

+++ src/app/requests/requests.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Services } from '../services';
+import { MatDialog } from '@angular/material';
 import { AuthServices } from '../auth.service';
-
+import { MatDialogComponent } from '../mat-dialog/mat-dialog.component';
 @Component({
   selector: 'app-requests',
   templateUrl: './requests.component.html',
@@ -12,7 +13,11 @@ import { AuthServices } from '../auth.service';
 export class RequestsComponent implements OnInit {
   requestsData:any;
   loading = true;
-  constructor(private AuthServices:AuthServices,public Services:Services, private router : Router, private route:ActivatedRoute) {
+  constructor(private AuthServices:AuthServices,
+              public Services:Services, 
+              private router : Router, 
+              private route:ActivatedRoute,
+              private dialog: MatDialog) {
     this.AuthServices.userAuth('requests');
   }
 
@@ -26,9 +31,28 @@ export class RequestsComponent implements OnInit {
      this.loading = false;
     });
   }
-  rejectRequest(id,index){
-    this.Services.deleteRequests(id).subscribe(result=>{
-      this.requestsData.splice(index,1);
+  rejectRequest(id){
+    var modelData = {title:"CONFIRM",message:"Are you sure want to delete?", modelType:"confirm"};
+    this.openDialog(modelData,id);
+  }
+
+  openDialog(content,id){
+    var self = this;
+    const dialogRef = this.dialog.open(MatDialogComponent, {
+      height: '250px',
+      width: '300px',
+      data: content
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed'+result);
+      if(result == 'Confirm'){
+        this.Services.deleteRequests(id).subscribe(result=>{
+        var modelData = {title:"SUCCESS",message:"Request Deleted Successfully", modelType:"default"};
+        this.openDialog(modelData,"");  
+        this.getRequestDetails();
+    });
+      }
     });
   }
 }
